+++ CHANGELOG.md
@@ -6,6 +6,7 @@ master
 
 * Switch to the new security checker
 * Migrate from Travis to GitHub Actions
+* Fix deprecated PHP-CS rules
 
 v0.2.0
 ------